requirements.txt
streamlit>=1.28.0
opencv-python-headless
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe>=0.10.0
imageio>=2.31.0
transformers>=4.30.0
scipy>=1.11.0
svgwrite>=1.4.3